iap/requirements.txt
Flask==0.12.2
Werkzeug<0.13.0,>=0.12.0
google-api-python-client
requests
httplib2
oauth2client
pyopenssl